fixed overflow bug, where high framerates would drastically drop the number of displayed points
git-svn-id: svn://svn.code.sf.net/p/slam6d/code/trunk@322 51f1e6e7-7289-49ad-8748-f09e43fc6daa

+++ src/show/show_gl.cc
@@ -7,7 +7,7 @@ bool fullydisplayed = true;   // true if all points have been drawn to the scree
 bool showall = false;         // true iff next DrawPoints should redraw scene with all points
 bool mousemoving = false;     // true iff a mouse button has been pressed inside a window, but hs not been released
 bool delayeddisplay = false;  // true iff mouse button callbacks should redraw the scene after button release
-long ptstodisplay = 100000;  
+double ptstodisplay = 100000;  
 double lastfps = idealfps;    // last frame rate    
 int pointmode = -1;
